+++ CHANGELOG.md
@@ -9,3 +9,4 @@
     - Skips execution in the admin panel unless it’s an AJAX request.
     - Ensures that WooCommerce is fully initialized before running.
     - If the cart is empty, it returns nothing instead of causing errors.
+- Plugin settings page enhancement